* Refactor and skeleton finished
