+++ pallets/ajuna-awesome-avatars/src/types/avatar/versions/v4/constants.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+pub(crate) const MIN_SACRIFICE: usize = 1;
+pub(crate) const MAX_SACRIFICE: usize = 4;
